Make repeated runs of same model give same outputs.
Without this change the naive "greedy first match" algorithm we use to
match gapped billets to an available Sailor may result in different
numbers of billets being filled depending on the order the billets are
looked at. This makes repeated runs of the model hard to compare
against strategy changes.

Fix by forcing the list of billets to fill to be sorted first.